Move 1: Split to cmd and pkg. Testing: start controller from go
