+++ .env-template-azurerm
@@ -4,4 +4,4 @@ ARM_SUBSCRIPTION_ID=
 ARM_TENANT_ID=
 PROVIDER_CONFIG_HCL=features {}
 TF_PROVIDER_NAME=azurerm
-TF_PROVIDER_PATH=./hack/.terraform/plugins/linux_amd64/
+TF_PROVIDER_PATH=./../../hack/.terraform/plugins/linux_amd64/
