+++ cmd/server/integration_test.go
@@ -1,4 +1,4 @@
-package main_test
+package main
 
 import (
 	"context"
