+++ cmd/server/main.go
@@ -4,7 +4,10 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/lawrencegripper/tfoperatorbridge/tfprovider"
+	"github.com/go-logr/logr"
+	"github.com/go-openapi/spec"
+	"github.com/lawrencegripper/tfoperatorbridge/pkg/k8s"
+	"github.com/lawrencegripper/tfoperatorbridge/pkg/tfprovider"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -13,6 +16,13 @@ import (
 func main() {
 	log := ctrl.Log.WithName("main")
 
+	provider, schemas, resources := setupCRDs(log)
+
+	// Start an informer to watch for crd items
+	k8s.StartControllerRuntime(provider, resources, schemas)
+}
+
+func setupCRDs(log logr.Logger) (*tfprovider.TerraformProvider, []spec.Schema, []k8s.GroupVersionFull) {
 	// Get a provider instance by installing or using existing binary
 	provider, err := tfprovider.SetupProvider(log)
 	if err != nil {
@@ -20,13 +30,12 @@ func main() {
 	}
 
 	// Creating CRDs in K8s with correct structure based on TF Schemas
-	resources, schemas, err := createK8sCRDsFromTerraformProvider(provider)
+	resources, schemas, err := k8s.CreateK8sCRDsFromTerraformProvider(getK8sClientConfig(), provider)
 	if err != nil {
 		panic(err)
 	}
 
-	// Start an informer to watch for crd items
-	setupControllerRuntime(provider, resources, schemas)
+	return provider, schemas, resources
 }
 
 func homeDir() string {
