+++ pkg/k8s/cipher.go
@@ -1,4 +1,4 @@
-package main
+package k8s
 
 import (
 	"bytes"
