+++ pkg/k8s/controller.go
@@ -1,4 +1,4 @@
-package main
+package k8s
 
 import (
 	"context"
@@ -10,7 +10,7 @@ import (
 
 	openapi_spec "github.com/go-openapi/spec"
 	"github.com/hashicorp/terraform/providers"
-	"github.com/lawrencegripper/tfoperatorbridge/tfprovider"
+	"github.com/lawrencegripper/tfoperatorbridge/pkg/tfprovider"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -86,7 +86,8 @@ func runtimeObjFromGVK(r schema.GroupVersionKind) runtime.Object {
 	return obj
 }
 
-func setupControllerRuntime(provider *tfprovider.TerraformProvider, resources []GroupVersionFull, schemas []openapi_spec.Schema) {
+// StartControllerRuntime starts the validating webhook and the reconciliation loops
+func StartControllerRuntime(provider *tfprovider.TerraformProvider, resources []GroupVersionFull, schemas []openapi_spec.Schema) {
 	ctrl.SetLogger(zap.Logger(true))
 
 	mgr, err := ctrl.NewManager(ctrl.GetConfigOrDie(), ctrl.Options{})
